Split functions in the step function into .hpp and .cpp files
This change should improve compile times, especially with incremental builds
